fixing liking and following

+++ handlers/favorites.go
@@ -6,6 +6,7 @@ import (
 	"log"
 	"net/http"
 
+	"github.com/Edafi/MusicVibe/middleware"
 	"github.com/Edafi/MusicVibe/models"
 	"github.com/gorilla/mux"
 )
@@ -16,7 +17,13 @@ type FavoritesHandler struct {
 
 // Получить избранные треки
 func (handler *FavoritesHandler) GetFavoriteTracks(response http.ResponseWriter, request *http.Request) {
-	userID := request.Context().Value("id").(string)
+	val := request.Context().Value(middleware.ContextUserIDKey)
+	userID, ok := val.(string)
+	if !ok {
+		log.Println("UserID not found in context")
+		http.Error(response, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
 
 	query := `
 		SELECT t.id, t.title, m.id, m.name, t.image_path, t.audio_path,
@@ -52,7 +59,14 @@ func (handler *FavoritesHandler) GetFavoriteTracks(response http.ResponseWriter,
 
 // Добавить трек в избранное
 func (handler *FavoritesHandler) AddFavoriteTrack(response http.ResponseWriter, request *http.Request) {
-	userID := request.Context().Value("id").(string)
+	val := request.Context().Value(middleware.ContextUserIDKey)
+	userID, ok := val.(string)
+	if !ok {
+		log.Println("UserID not found in context")
+		http.Error(response, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	trackID := mux.Vars(request)["trackId"]
 
 	_, err := handler.DB.Exec(`INSERT IGNORE INTO liked_tracks (user_id, track_id) VALUES (?, ?)`, userID, trackID)
@@ -67,7 +81,14 @@ func (handler *FavoritesHandler) AddFavoriteTrack(response http.ResponseWriter, 
 
 // Удалить трек из избранного
 func (handler *FavoritesHandler) DeleteFavoriteTrack(response http.ResponseWriter, request *http.Request) {
-	userID := request.Context().Value("id").(string)
+	val := request.Context().Value(middleware.ContextUserIDKey)
+	userID, ok := val.(string)
+	if !ok {
+		log.Println("UserID not found in context")
+		http.Error(response, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	trackID := mux.Vars(request)["trackId"]
 
 	_, err := handler.DB.Exec(`DELETE FROM liked_tracks WHERE user_id = ? AND track_id = ?`, userID, trackID)
@@ -82,7 +103,13 @@ func (handler *FavoritesHandler) DeleteFavoriteTrack(response http.ResponseWrite
 
 // Получить избранные альбомы
 func (handler *FavoritesHandler) GetFavoriteAlbums(response http.ResponseWriter, request *http.Request) {
-	userID := request.Context().Value("id").(string)
+	val := request.Context().Value(middleware.ContextUserIDKey)
+	userID, ok := val.(string)
+	if !ok {
+		log.Println("UserID not found in context")
+		http.Error(response, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
 
 	query := `
 		SELECT a.id, a.title, a.image_path, YEAR(a.release_date), m.id, m.name
@@ -118,7 +145,14 @@ func (handler *FavoritesHandler) GetFavoriteAlbums(response http.ResponseWriter,
 
 // Добавить альбом в избранное
 func (handler *FavoritesHandler) AddFavoriteAlbum(response http.ResponseWriter, request *http.Request) {
-	userID := request.Context().Value("id").(string)
+	val := request.Context().Value(middleware.ContextUserIDKey)
+	userID, ok := val.(string)
+	if !ok {
+		log.Println("UserID not found in context")
+		http.Error(response, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	albumID := mux.Vars(request)["albumId"]
 
 	_, err := handler.DB.Exec(`INSERT IGNORE INTO liked_albums (user_id, album_id) VALUES (?, ?)`, userID, albumID)
@@ -133,7 +167,14 @@ func (handler *FavoritesHandler) AddFavoriteAlbum(response http.ResponseWriter, 
 
 // Удалить альбом из избранного
 func (handler *FavoritesHandler) DeleteFavoriteAlbum(response http.ResponseWriter, request *http.Request) {
-	userID := request.Context().Value("id").(string)
+	val := request.Context().Value(middleware.ContextUserIDKey)
+	userID, ok := val.(string)
+	if !ok {
+		log.Println("UserID not found in context")
+		http.Error(response, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	albumID := mux.Vars(request)["albumId"]
 
 	_, err := handler.DB.Exec(`DELETE FROM liked_albums WHERE user_id = ? AND album_id = ?`, userID, albumID)
